Freeze time, share fixture
